+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - using baseCommit when rebasing the worktree on different branch
 - added base branch name to the worktree merge button and improved rebasing
 - hide save prompt button when editing last message
+- added support for redoing any user prompt
 
 ## [0.63.0]